修复：中断功能

+++ common/function_subprocess.py
@@ -4,4 +4,5 @@ import subprocess
 def stop_process(process: subprocess.Popen):
     """终止调用"""
     print('中断调用')
-    process.terminate()
+    if process:
+        process.terminate()
